test/requirements.txt
anyio==4.4.0
asgiref==3.8.1
Authlib==1.3.1
bleak==0.22.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclasses==0.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-secret-manager==2.20.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lmcloud==1.1.13
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
requests==2.32.3
rsa==4.9
sniffio==1.3.1
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
